server/requirements.txt
fastapi
uvicorn[standard]
asyncpg
paho-mqtt
python-dotenv
httpx
pydantic
openai
